# Conflicts:
#	mooc-authorize2/src/main/java/cn/edu/lingnan/authorize/aspect/LoginLogAspect.java
#	mooc-authorize2/src/main/java/cn/edu/lingnan/authorize/config/WebMvcConfig.java
#	mooc-core/src/main/java/cn/edu/lingnan/core/authentication/config/InterceptorConfig.java

+++ mooc-portal/src/main/java/cn/edu/lingnan/mooc/portal/dao/CollectionRepository.java
@@ -10,5 +10,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
  * @data 2021/02/02
  * 收藏表repository
  */
-public interface CollectionRepository extends JpaRepository<Collection, Integer>, JpaSpecificationExecutor<Collection> {
+public interface CollectionRepository extends JpaRepository<Collection, Long>, JpaSpecificationExecutor<Collection> {
 }
