+++ mooc-portal/src/main/java/cn/edu/lingnan/mooc/portal/model/entity/MonitorRecord.java
@@ -15,21 +15,35 @@ import java.util.Date;
 @DynamicInsert
 @Entity(name = "course_monitor_record")
 public class MonitorRecord {
-    // 主键   
+    /**
+     * 主键
+     */
     @Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-    private Integer id;
-    // 管理员账号   
-    private Integer teacherId;
-    // 课程id   
-    private Integer courseId;
-    // 具体消息   
+    @GeneratedValue(strategy=GenerationType.IDENTITY)
+    private Long id;
+    /**
+     * 教师id
+     */
+    private Long teacherId;
+    /**
+     * 课程id
+     */
+    private Long courseId;
+    /**
+     * 具体消息
+     */
     private String message;
-    // 类型|新增课程、上传视频、删除课程、课程点赞、取消点赞
+    /**
+     * 类型|新增课程、上传视频、删除课程
+     */
     private String recordType;
-    // 登录ip   
+    /**
+     * 登录ip
+     */
     private String ip;
-    // 创建时间   
+    /**
+     * 创建时间
+     */
     private Date createTime;
     
 }
