+++ mooc-portal/src/main/java/cn/edu/lingnan/mooc/portal/model/vo/SectionVO.java
@@ -12,24 +12,42 @@ import java.util.Date;
 @Data
 @ToString
 public class SectionVO {
-    private Integer id;
-    // 标题
+    private Long id;
+    /**
+     * 标题
+     */
     private String title;
-    // 课程id
-    private Integer courseId;
-    // 章节id
-    private Integer chapterId;
-    // 视频
+    /**
+     *
+     */
+    private Long courseId;
+    /**
+     * 章节id
+     */
+    private Long chapterId;
+    /**
+     * 视频
+     */
     private String video;
-    // 对应文件表ID
+    /**
+     * 对应文件表ID
+     */
     private Integer fileId;
-    // 时长|单位秒
+    /**
+     * 时长|单位秒
+     */
     private Integer duration;
-    // 顺序
+    /**
+     * 顺序
+     */
     private Integer sort;
-    // 创建时间
+    /**
+     * 创建时间
+     */
     private Date createTime;
-    // 修改时间
+    /**
+     * 修改时间
+     */
     private Date updateTime;
